requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# OpenAI
openai==1.3.0

# Database
pymysql==1.1.0
DBUtils==3.1.0
aiomysql==0.2.0

# Vector Search
qdrant-client==1.7.0
numpy==1.24.3

# Elasticsearch
elasticsearch==8.11.0

# Embedding
sentence-transformers==2.7.0
torch==2.3.0

# Utilities
python-multipart==0.0.6
tenacity==8.2.3
tiktoken==0.5.2
httpx==0.25.2
aiofiles==23.2.1
